Restructured component folders, made UI responsive
