--- src/components/Table.module.css
@@ -1,22 +1,0 @@
-.table {
-  margin: 0;
-  width: 100%;
-}
-
-.table caption {
-  font-size: 1.5rem;
-  font-weight: 600;
-  text-align: start;
-  margin-bottom: 1rem;
-}
-
-.table th {
-  text-align: left;
-  padding-bottom: 0.5rem;
-}
-
-.table td {
-  padding-right: 1rem;
-  padding-top: 0.5rem;
-  padding-bottom: 0.5rem;
-}
