+++ src/components/Table/Table.module.css
@@ -1,0 +1,56 @@
+.table {
+  margin: 0;
+  width: 100%;
+}
+
+.table caption {
+  font-size: 1.5rem;
+  font-weight: 600;
+  text-align: start;
+  margin-bottom: 1rem;
+}
+
+.table th {
+  text-align: left;
+  padding-bottom: 0.5rem;
+}
+
+.table td {
+  padding-right: 1rem;
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem;
+  text-overflow: ellipsis;
+  overflow: hidden;
+}
+
+/* Tablet and Smaller Desktop */
+@media (min-width: 740px) and (max-width: 1120px) {
+  .table {
+    table-layout: fixed;
+  }
+  .table caption {
+    font-size: 1.25rem;
+  }
+  .table tr th:nth-child(1) {
+    width: 50%;
+  }
+  .table tr th:nth-child(2) {
+    width: 25%;
+  }
+}
+
+/* Mobile */
+@media (max-width: 739px) {
+  .table {
+    table-layout: fixed;
+  }
+  .table caption {
+    font-size: 1rem;
+  }
+  .table tr th:nth-child(1) {
+    width: 50%;
+  }
+  .table tr th:nth-child(2) {
+    width: 25%;
+  }
+}
